Refactor OrderInterest controller and service.

+++ FREELA-API/src/main/java/freela/api/FREELAAPI/domain/services/impl/OrderInterrestServiceImpl.java
@@ -1,6 +1,7 @@
 package freela.api.FREELAAPI.domain.services.impl;
 
 import freela.api.FREELAAPI.application.web.helpers.ListaObj;
+import freela.api.FREELAAPI.domain.exceptions.DataAccessException;
 import freela.api.FREELAAPI.domain.repositories.OrderInterestRepository;
 import freela.api.FREELAAPI.domain.repositories.OrderRepository;
 import freela.api.FREELAAPI.domain.repositories.SubCategoryRepository;
@@ -9,6 +10,7 @@ import freela.api.FREELAAPI.resourses.entities.OrderInterest;
 import freela.api.FREELAAPI.resourses.entities.Orders;
 import freela.api.FREELAAPI.resourses.entities.SubCategory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
@@ -18,42 +20,42 @@ import java.util.Optional;
 @Service
 public class OrderInterrestServiceImpl implements OrderInterrestService {
 
-    @Autowired
-    private OrderRepository orderRepository;
-    @Autowired
-    private OrderInterestRepository orderInterestRepository;
-    @Autowired
-    private SubCategoryRepository subCategoryRepository;
+    private final OrderRepository orderRepository;
+    private final OrderInterestRepository orderInterestRepository;
+    private final SubCategoryRepository subCategoryRepository;
+
+    public OrderInterrestServiceImpl(
+            OrderRepository orderRepository,
+            OrderInterestRepository orderInterestRepository,
+            SubCategoryRepository subCategoryRepository
+    ) {
+        this.orderRepository = orderRepository;
+        this.orderInterestRepository = orderInterestRepository;
+        this.subCategoryRepository = subCategoryRepository;
+    }
 
     @Override
-
     public ListaObj<SubCategory> findByOrder(Integer id) {
-        try {
-            Optional<Orders> order = this.orderRepository.findById(id);
-            List<OrderInterest> interests =  this.orderInterestRepository.findAllByOrder(order.get());
-            ListaObj<SubCategory> subCategoryListaObj = new ListaObj<>(interests.size());
+        Orders order = findOrderById(id);
+        List<OrderInterest> interests = orderInterestRepository.findAllByOrder(order);
 
-            for(OrderInterest orderInterest : interests){
-                subCategoryListaObj.adiciona(orderInterest.getSubCategory());
-            }
-
-
-              return subCategoryListaObj;
-        }catch (RuntimeException ex){
-            throw new RuntimeException(ex.getMessage());
+        if (interests.isEmpty()) {
+            throw new DataAccessException("Lista de pedidos vazia.", HttpStatus.NOT_FOUND);
         }
+
+        ListaObj<SubCategory> subCategoryListaObj = new ListaObj<>(interests.size());
+
+        interests.stream()
+                .map(OrderInterest::getSubCategory)
+                .forEach(subCategoryListaObj::adiciona);
+
+        return subCategoryListaObj;
     }
-    public void createOrderInterest(ArrayList<Integer> subCategories, Orders order){
-
-        for(Integer subCategoryid : subCategories){
-            Optional<SubCategory> subCategory = this.subCategoryRepository.findById(subCategoryid);
-            subCategory.ifPresent(category -> this.orderInterestRepository.save(
-                    new OrderInterest(
-                            order,
-                            category
-                    )
-            ));
-        }
+    public void createOrderInterest(ArrayList<Integer> subCategories, Orders order) {
+        subCategories.stream()
+                .map(subCategoryRepository::findById)
+                .flatMap(Optional::stream)
+                .forEach(subCategory -> orderInterestRepository.save(new OrderInterest(order, subCategory)));
     }
 
     public void updateOrderInterest(ArrayList<Integer> subCategories, Orders order){
@@ -63,6 +65,21 @@ public class OrderInterrestServiceImpl implements OrderInterrestService {
 
     public void deleteOrderInterest(Orders orders){
         List<OrderInterest> interests = this.orderInterestRepository.findAllByOrder(orders);
+
+        if (interests.isEmpty()){
+            throw new DataAccessException("Nenhum interesse encontrado.", HttpStatus.NOT_FOUND);
+        }
+
         this.orderInterestRepository.deleteAll(interests);
+    }
+
+    private Orders findOrderById(Integer orderId) {
+        Optional<Orders> order = this.orderRepository.findById(orderId);
+        if (order.isEmpty()) {
+            throw new DataAccessException("Pedido não encontrado.", HttpStatus.NOT_FOUND);
+        } else if (order.get().isAccepted()) {
+            throw new DataAccessException("Pedido já foi aceito.", HttpStatus.BAD_REQUEST);
+        }
+        return order.get();
     }
 }
